// Umbraco.AI/src/Umbraco.AI.Web.StaticAssets/Client/src/core/styles/code-block.styles.ts
import { css } from "@umbraco-cms/backoffice/external/lit";

export const codeBlockStyles = css`
    .code-block {
        background: var(--uui-color-surface-alt);
        padding: 12px;
        border-radius: 4px;
        overflow-x: auto;
        font-family: monospace;
        font-size: 12px;
        margin: 0;
        white-space: pre-wrap;
        word-break: break-word;
    }
`;